requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
Django==5.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-htmx==1.21.0
django-password-validators==1.7.3
django-use-email-as-username==1.4.0
django-widget-tweaks==1.5.0
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==24.1
pandocfilters==1.5.1
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.6
pure_eval==0.2.3
pydot==3.0.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
